requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
docopt==0.6.2
docutils==0.21.2
Flask==3.1.0
fonttools==4.57.0
gunicorn==23.0.0
id==1.5.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
joblib==1.4.2
keyring==25.6.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplot==0.1.9
matplotlib==3.10.1
mdurl==0.1.2
more-itertools==10.7.0
nh3==0.2.21
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pipreqs==0.4.13
Pygments==2.19.1
pyloco==0.0.139
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
SimpleWebSocketServer==0.1.2
six==1.17.0
sklearn-compat==0.1.3
threadpoolctl==3.6.0
twine==6.1.0
typing==3.7.4.3
tzdata==2025.2
urllib3==2.4.0
ushlex==0.99.1
waitress==3.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
yarg==0.1.10
